requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0     # via wagtail
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
dj-database-url==0.5.0
django-modelcluster==4.1  # via wagtail
django-taggit==0.22.2     # via wagtail
django-treebeard==4.3     # via wagtail
django==2.0.5
djangorestframework==3.8.2  # via wagtail
draftjs-exporter==2.0.0   # via wagtail
first==2.0.1              # via pip-tools
html5lib==0.999999999     # via wagtail
idna==2.6                 # via requests
olefile==0.43
pillow==4.3.0
pip-tools==2.0.2
psycopg2-binary==2.7.4
pytz==2018.4
requests==2.18.4          # via wagtail
six==1.11.0               # via html5lib, pip-tools
unidecode==0.4.21         # via wagtail
urllib3==1.22             # via requests
wagtail==2.0.1
webencodings==0.5.1       # via html5lib
willow==1.1               # via wagtail
